+++ engine/target/wasm32-unknown-unknown/release/draftr.d
@@ -1,0 +1,1 @@
+D:\Joshua\Projects\draftr.design\engine\target\wasm32-unknown-unknown\release\draftr.wasm: D:\Joshua\Projects\draftr.design\engine\src\lib.rs